+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Concurrency limiter dependency (in a similar style to cooldowns).
 - `disable_bucket` method to the in-memory concurrency and cooldown manager impls.
+- `any_checks`/`with_any_checks` and `all_checks`/`with_all_checks` functions for more garnular check
+  flow control. `any_checks` passes if any of the provided checks pass and `all_checks` passes if all
+  the provided checks pass while both ensure the checks are run sequentially rather than concurrently.
 
 ### Changed
 - `cached_inject` and `cache_callback` now both accept `float` and `int` seconds for `expire_after`.